adding inheritance base planner
