+++ configs/frenet_optimal_trajectory_planner.conf
@@ -3,22 +3,21 @@
 --carla_traffic_light_detection
 ######### Prediction config #########
 --perfect_obstacle_tracking
---tracking_num_steps=10
+--tracking_num_steps=5
 --prediction
 --prediction_type=linear
---prediction_num_past_steps=10
---prediction_num_future_steps=10
+--prediction_num_past_steps=5
+--prediction_num_future_steps=2
 ######### Planning config #########
 --planning_type=frenet_optimal_trajectory
---target_speed=17.5
---obstacle_radius=3.0
+--target_speed=20
+--obstacle_radius=2.75
 ######### Control config #########
 --control_agent=pid
 --stop_for_vehicles=False
 --stop_for_people=False
 --stop_for_traffic_lights=False
 --pid_steer_wp=4
---steer_gain=0.6
 --pid_speed_wp=-1
 ######### Logging config #########
 --log_file_name=pylot.log
