+++ configs/rrt_star_planner.conf
@@ -16,7 +16,7 @@
 --stop_for_people=False
 --stop_for_traffic_lights=False
 --target_speed=20
---pid_steer_wp=5
+--pid_steer_wp=2
 --pid_speed_wp=-1
 ######### Logging config #########
 --log_file_name=pylot.log
